+++ tests/unit/test_session.py
@@ -31,6 +31,8 @@ from tests.conftest import (
     MOCKED_URL_REDIRECT,
     MOCKED_URL_REDIRECT_TARGET,
     MOCKED_URL_VARY,
+    MOCKED_URL_VARY_REDIRECT,
+    MOCKED_URL_VARY_REDIRECT_TARGET,
     START_DT,
     ignore_deprecation,
     patch_normalize_url,
@@ -330,15 +332,39 @@ def test_match_headers__vary(mock_session):
     """Vary should be used to validate headers, if available.
     It should also override `match_headers` for the secondary cache key, if both are provided.
     """
-    # mock_session.settings.match_headers = ['Accept-Encoding']
     headers_1 = {'Accept': 'application/json', 'User-Agent': 'qutebrowser'}
     headers_2 = {'Accept': 'application/json', 'User-Agent': 'Firefox'}
     headers_3 = {'Accept': 'text/plain', 'User-Agent': 'qutebrowser'}
+    mock_session.get(MOCKED_URL_VARY, headers=headers_1)
 
-    assert mock_session.get(MOCKED_URL_VARY, headers=headers_1).from_cache is False
     assert mock_session.get(MOCKED_URL_VARY, headers=headers_1).from_cache is True
     assert mock_session.get(MOCKED_URL_VARY, headers=headers_2).from_cache is True
     assert mock_session.get(MOCKED_URL_VARY, headers=headers_3).from_cache is False
+
+
+@pytest.mark.parametrize(
+    'headers, expected_from_cache',
+    [
+        ({'Accept': 'text/plain', 'Accept-Language': 'en-US'}, True),
+        ({'Accept-Language': 'en-US'}, True),
+        ({'Accept-Language': 'en-GB'}, False),
+        ({'Accept': 'application/json'}, False),
+        ({}, False),
+    ],
+)
+def test_match_headers__vary_with_redirects(headers, expected_from_cache, mock_session):
+    """With Vary + redirects, Vary should match against the last request in the redirect chain.
+    * 1st response has `Vary: Accept` (shouldn't be matched)
+    * 2nd response has `Vary: Accept-Language` (should be matched)
+    """
+    r = mock_session.get(
+        MOCKED_URL_VARY_REDIRECT,
+        headers={'Accept': 'text/plain', 'Accept-Language': 'en-US'},
+    )
+    assert r.url == MOCKED_URL_VARY_REDIRECT_TARGET
+
+    r2 = mock_session.get(MOCKED_URL_VARY_REDIRECT, headers=headers)
+    assert r2.from_cache is expected_from_cache
 
 
 def test_include_get_headers():